Deleted AdminService, SellerService, VehicleController and moved the method in the correct files
