+++ src/main/java/com/develhope/spring/controllers/AdminController.java
@@ -1,8 +1,7 @@
-package com.develhope.spring.controllers.users;
+package com.develhope.spring.controllers;
 
-import com.develhope.spring.entities.operations.Order;
 import com.develhope.spring.entities.vehicle.Vehicle;
-import com.develhope.spring.services.users.AdminService;
+import com.develhope.spring.services.VehicleService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -12,30 +11,34 @@ import java.util.List;
 @RequestMapping("/admin")
 public class AdminController {
     @Autowired
-    private AdminService adminService;
+    private VehicleService vehicleService;
 
     @PostMapping("/vehicle/create")
     public Vehicle createVehicle(@RequestBody Vehicle vehicle) {
-        return adminService.createVehicle(vehicle);
+        return vehicleService.createVehicle(vehicle);
     }
 
     @PutMapping("/vehicle/update/{id}")
     public Vehicle updateVehicle(@PathVariable long id, @RequestBody Vehicle vehicle) {
-        return adminService.updateVehicle(id, vehicle);
+        return vehicleService.updateVehicle(id, vehicle);
     }
 
-    @DeleteMapping("vehicle/delete/{id}")
+    @DeleteMapping("/vehicle/delete/{id}")
     public Boolean deleteVehicle(@PathVariable long id) {
-        return adminService.deleteVehicle(id);
+        return vehicleService.deleteVehicle(id);
     }
 
-    @PatchMapping("vehicle/status/update/{id}")
+    @PatchMapping("/vehicle/status/update/{id}")
     public Vehicle updateVehicleStatusFromId(@PathVariable long id, @RequestParam String status) {
-        return adminService.updateVehicleStatusFromId(id, status);
+        return vehicleService.updateVehicleStatusFromId(id, status);
     }
 
-    @GetMapping("vehicle/get/by/status/used")
+    @GetMapping("/vehicle/get/by/status/used")
     public List<Vehicle> findByStatusAndUsed(@RequestParam String status, @RequestParam Boolean used) {
-        return adminService.findByStatusAndUsed(status, used);
+        return vehicleService.findByStatusAndUsed(status, used);
+    }
+    @GetMapping("/vehicles/get")
+    public List<Vehicle> getAllVehicles(){
+        return vehicleService.getAllVehicles();
     }
 }
